docs: comments-write 파일용도에 맞게 분기함
